Refactor offset example some more
