+++ src/types/Middleware.ts
@@ -1,5 +1,5 @@
-import { GraphQLFieldConfig, GraphQLObjectType } from 'graphql'
-import { ObjectConfig, FieldConfig } from './ObjectConfig'
+import { GraphQLFieldConfig, GraphQLObjectType, GraphQLFieldResolver } from 'graphql'
+import { ObjectConfig, FieldConfig, SalesforceObjectConfig } from './ObjectConfig'
 
 export type BuildObjectsMiddleware = (
    field: GraphQLFieldConfig<any, any> & Readonly<FieldConfig>,
@@ -7,3 +7,7 @@ export type BuildObjectsMiddleware = (
    parent: Readonly<ObjectConfig>,
    objectMap: { readonly [name: string]: GraphQLObjectType }
   ) => GraphQLFieldConfig<any, any>
+
+export type ResolverMiddleware =
+  (rootQuery: SalesforceObjectConfig,
+   objectMap: { [x: string]: SalesforceObjectConfig }) => GraphQLFieldResolver<any, any>
